kardex valorado terminado

+++ database/migrations/2019_05_02_161345_create_article_request_items_table.php
@@ -19,8 +19,8 @@ class CreateArticleRequestItemsTable extends Migration
             $table->foreign('article_request_id')->references('id')->on('article_requests');
             $table->integer('article_id');
             $table->foreign('article_id')->references('id')->on('articles');
-            $table->integer('quantity');
-            $table->integer('quantity_apro');
+            $table->decimal('quantity');
+            $table->decimal('quantity_apro');
             $table->integer('article_income_item_id')->nullable(); //entradas
             $table->foreign('article_income_item_id')->references('id')->on('article_income_items');
             $table->timestamps();
